If You run a filter against the live service to only show certain status codes and there are no results for that filter, PYWB will attempt to ditch your filter and just return the first 100 results it has for that URL.
So, by default PYWB will try to get all 2xx and 3xx results.  If none exist, it will then give you 100 4xx and 5xx results it has for that URL.

I've changed the logic to not do the extra return if filters are supplied to the query. I need this to be tested extensively and confirmed non-destructive before we merge.